unites readable and writable ColumnIndex and RowIndex for page breaks
