+++ src/jxl/write/biff/VerticalPageBreaksRecord.java
@@ -30,35 +30,6 @@ import jxl.read.biff.IVerticalPageBreaks;
  */
 class VerticalPageBreaksRecord extends WritableRecordData implements IVerticalPageBreaks {
   
-  public static class ColumnIndex {
-    private final int firstColumnFollowingBreak;
-    private final int firstRow;
-    private final int lastRow;
-    
-    public ColumnIndex(int firstRowBelowBreak, int firstRow, int lastRow) {
-      this.firstColumnFollowingBreak = firstRowBelowBreak;
-      this.firstRow = firstRow;
-      this.lastRow = lastRow;
-    }
-
-    public int getFirstColumnFollowingBreak() {
-      return firstColumnFollowingBreak;
-    }
-
-    public int getFirstRow() {
-      return firstRow;
-    }
-
-    public int getLastRow() {
-      return lastRow;
-    }
-
-    private ColumnIndex withFirstColumnFollowingBreak(int i) {
-      return new ColumnIndex(i, firstRow, lastRow);
-    }
-    
-  }
-
   /**
    * The row breaks
    */
